スクリプトの大幅修正
・変更点
・Playerクラスを分解し、HPクラス・TRICKPointクラス・FEVERPointクラスに分けました
・Enemyクラスを消し、代わりにHPクラスをアタッチしました
・ManagementOfScoreクラスのスコア算出処理とプレイヤーの被弾回数の処理を一部改造しました
・StatusDisplayクラスを分解し、HpDisplayクラス・TrickPointDisplayクラス・FeverPointDisplayクラスに分解して敵とプレイヤーのHPのゲージ処理両方をHpDisplayクラスでできるようにしました
・桑原君が前に書いてくれたバフストックのゲージ処理はBuffStockDisplayクラスに移しておいたので、先生に成果物(バフストックのゲージ処理部分のスクリプト)を見せる時はそれを見せてください
