+++ Big Wave prototype/Assets/Script/EtcScript/SceneControlManager.cs
@@ -18,16 +18,6 @@ public class SceneControlManager : MonoBehaviour
         
     }
 
-    public void ChangeGameoverScene()//ゲームオーバー画面に移行
-    {
-        SceneManager.LoadScene("GameoverScene");
-    }
-
-    public void ChangeClearScene()//クリア画面に移行
-    {
-        SceneManager.LoadScene("ClearScene");
-    }
-
     public void ChangeMenuScene()//メニュー画面に移行
     {
         SceneManager.LoadScene("MenuScene");
