+++ Big Wave prototype/Assets/Script/GamePosScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa9112701627fd44d99ff58da5bf9d8f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
