+++ Big Wave prototype/Assets/Script/NoNeededScript/ControlGamespeed.cs
