+++ Big Wave prototype/Assets/Script/NoNeededScript/ScoreDisplay.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 
 public class ScoreDisplay : MonoBehaviour
 {
-    [Header("表示させるテキスト")]
-    [SerializeField] TMP_Text Score_UI;//表示させるテキスト
+    //[Header("表示させるテキスト")]
+    //[SerializeField] TMP_Text Score_UI;//表示させるテキスト
 
     // Start is called before the first frame update
     void Start()
@@ -17,6 +17,6 @@ public class ScoreDisplay : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        Score_UI.text = "Score: " + ManagementOfScore.NowScore.ToString("");
+        //Score_UI.text = "Score: " + ManagementOfScore.NowScore.ToString("");
     }
 }
