+++ Big Wave prototype/Assets/Script/NoNeededScript/ScoreDisplay.cs.meta
