+++ Big Wave prototype/Assets/Script/PlayerEnemyCommonScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9eb1ee306aac4e24cabf2880f25c5884
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
