+++ Big Wave prototype/Assets/Script/PlayerEnemyCommonScript/HP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d26110e1f4dbe240ac2caa7e101645d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
