+++ Big Wave prototype/Assets/Script/PlayerScript/FEVERPoint.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 37c6caf9150bfd14283d02dd47d6c15d
+guid: 8cc35e6b71177234ba12d439630e2be4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
