+++ Big Wave prototype/Assets/Script/PlayerScript/MoveControl.cs
@@ -2,16 +2,14 @@ using UnityEngine;
 
 public class MoveControl : MonoBehaviour
 {
-    //☆塩が書いた
+    //☆作成者:杉山
     [SerializeField] float speed = 11.5f;//移動スピード
     //[SerializeField] float trick_SpeedFactor=1f;//トリックをためた時のスピードの上昇具合、1、２、3、nだとそれぞれトリック満タン時、トリック空っぽの時のスピードの2、3、4、(1+1*n)倍になる
-    Player player;
 
 
     // Start is called before the first frame update
     void Start()
     {
-        player=gameObject.GetComponent<Player>();
     }
 
     // Update is called once per frame
