+++ Big Wave prototype/Assets/Script/PlayerScript/ProcessFeverMode.cs
@@ -20,7 +20,7 @@ public class ProcessFeverMode : MonoBehaviour
     private float currentChargeTrick_GrowthRate = 1f;//現在のフィーバー状態のチャージトリック量アップの増加率
     private bool feverNow=false;//今フィーバー状態か
 
-    Player player;
+    FEVERPoint player_FeverPoint;
 
     public float CurrentPowerUp_GrowthRate
     {
@@ -45,7 +45,7 @@ public class ProcessFeverMode : MonoBehaviour
         currentPowerUp_GrowthRate = 1f;
         currentChargeTrick_GrowthRate = 1f;
         feverNow = false;
-        player = gameObject.GetComponent<Player>();
+        player_FeverPoint = gameObject.GetComponent<FEVERPoint>();
     }
 
     // Update is called once per frame
@@ -63,14 +63,14 @@ public class ProcessFeverMode : MonoBehaviour
     {
         if (feverNow == false)
         {
-            player.FeverPoint += chargeFeverPoint[count-1];
+            player_FeverPoint.FeverPoint += chargeFeverPoint[count-1];
         }
     }
 
     //まだフィーバー状態になっていないかつフィーバーポイントが満タンになったらフィーバー状態に移行
     void ChangeFeverMode()
     {
-        if (feverNow == false && player.FeverPoint >= player.FeverPointMax)
+        if (feverNow == false && player_FeverPoint.FeverPoint >= player_FeverPoint.FeverPointMax)
         {
             feverNow = true;
             remainingFeverTime = feverTime;
@@ -102,7 +102,7 @@ public class ProcessFeverMode : MonoBehaviour
             currentChargeTrick_GrowthRate=chargeTrick_GrowthRate;
             feverEffect.SetActive(true);
             float ratio = remainingFeverTime / feverTime;
-            player.FeverPoint = player.FeverPointMax * ratio;
+            player_FeverPoint.FeverPoint = player_FeverPoint.FeverPointMax * ratio;
         }
         //フィーバー状態じゃない時は　
         //攻撃力とチャージトリック量が通常
