+++ Big Wave prototype/Assets/Script/PlayerScript/TRICKPoint.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 91d9d1173b809d548bba0cb4455b645a
+guid: dc1abf8c3a61ce541be7e55715f3598d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
