+++ Big Wave prototype/Assets/Script/UIScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3898076eeda6e7748bbd7c147d3734db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
