+++ Big Wave prototype/Assets/Script/UIScript/HpDisplay.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2234283f10fc27a49a8c477c86f7b6d6
+guid: 7279a53b74001fe468ad012d12e320fe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
