+++ Big Wave prototype/Assets/Script/UIScript/PauseControl.cs
