+++ Big Wave prototype/Assets/Script/UIScript/TimeDisplay.cs
